refactoring hdf5 links code
