+++ nexus_constructor/field_utils.py
@@ -7,8 +7,7 @@ from nexus_constructor.field_widget import FieldWidget
 from nexus_constructor.invalid_field_names import INVALID_FIELD_NAMES
 from nexus_constructor.model.component import Component
 from nexus_constructor.model.dataset import Dataset
-from nexus_constructor.model.link import Link
-from nexus_constructor.model.stream import StreamGroup
+from nexus_constructor.model.stream import Link, StreamGroup
 from nexus_constructor.validators import FieldType
 
 if TYPE_CHECKING:
@@ -87,6 +86,7 @@ def get_fields_with_update_functions(
             "EV42Stream",
             "F142Stream",
             "HS00Stream",
+            "Link",
         ],
         Callable,
     ]
